aventura_matematica_DjangoRest/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
importlib-metadata==6.8.0
Markdown==3.5.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
zipp==3.17.0
